+++ es6.js
@@ -9,35 +9,15 @@ module.exports = {
     "remitbri/es5",
     "remitbri/rules/ecmascript6",
   ],
+  parserOptions : {
+    // set to 3, 5 (default), 6, or 7 to specify the version of ECMAScript you want to use
+    ecmaVersion : 6,
+    ecmaFeatures : {
+      // enable support for the experimental object rest/spread properties
+      experimentalObjectRestSpread : true,
+    },
+  },
   env : {
     es6 : true,
-  },
-  ecmaFeatures : {
-    /*
-    **true by default in the es6 environment**
-
-    arrowFunctions : true, // enable arrow functions
-    binaryLiterals : true, // enable binary literals
-    blockBindings : true, // enable let and const (aka block bindings)
-    classes : true, // enable classes
-    defaultParams : true, // enable default function parameters
-    destructuring : true, // enable destructuring
-    forOf : true, // enable for-of loops
-    generators : true, // enable generators
-    newTarget: true
-    objectLiteralComputedProperties : true, // enable computed object literal property names
-    objectLiteralDuplicateProperties : true, // enable duplicate object literal properties in strict mode
-    objectLiteralShorthandMethods : true, // enable object literal shorthand methods
-    objectLiteralShorthandProperties : true, // enable object literal shorthand properties
-    octalLiterals : true, // enable octal literals
-    regexUFlag : true, // enable the regular expression u flag
-    regexYFlag : true, // enable the regular expression y flag
-    restParams : true, // enable the rest parameters
-    spread : true, // enable the spread operator for arrays
-    superInFunctions : true, // enable super references inside of functions
-    templateStrings : true, // enable template strings
-    unicodeCodePointEscapes : true, // enable code point escapes
-
-    */
   },
 }