+++ node.js
@@ -6,15 +6,8 @@
 "use strict"
 
 module.exports = {
-  extends : "./rules/nodeJS",
+  extends : "remitbri/rules/nodeJS",
   env : {
     node : true,
-  },
-  ecmaFeatures : {
-    /*
-    **true by default in the node environment**
-
-    globalReturn : true, // allow return statements in the global scope
-    */
   },
 }