+++ rules/nodeJS.js
@@ -9,22 +9,24 @@
 module.exports = {
   rules : {
     // Enforce Return After Callback
-    "callback-return" : [2, ["callback", "cb", "next"]],
-    // Disallow require() outside of the top-level module scope
-    "global-require" : 2,
+    "callback-return" : ["error", ["callback", "cb", "next"]],
+    //  Enforce require() on the top-level module scope
+    "global-require" : "error",
     // Enforce Callback Error Handling
-    "handle-callback-err" : [2, "^(err|error|anySpecificError)$"],
+    "handle-callback-err" : ["error", "^(err|error|anySpecificError)$"],
     // Disallow Mixed Requires
-    "no-mixed-requires" : [2, {grouping : false}],
+    "no-mixed-requires" : ["error", {grouping : false, allowCall : false}],
     // Disallow new require
-    "no-new-require" : 2,
-    // Disallow string concatenation when using _dirname and _filename
-    "no-path-concat" : 2,
+    "no-new-require" : "error",
+    // Disallow string concatenation when using `__dirname` and `__filename`
+    "no-path-concat" : "error",
+    // Disallow `process.env`
+    "no-process-env" : "error",
     // Disallow process.exit()
-    "no-process-exit" : 2,
+    "no-process-exit" : "error",
     // Disallow Node modules
-    "no-restricted-modules" : 0,
+    "no-restricted-modules" : "off",
     // Disallow Synchronous Methods
-    "no-sync" : 0,
+    "no-sync" : "off",
   },
 }